+++ js/store.js
@@ -1,19 +1,22 @@
+
 import { Product } from "./product.js";
- export const products = [
-  new Product(
-    "image/product1.jpg",
-    "Крем увлажняющий",
-    "Уход",
-    "Увлажнение и питание кожи",
-    ["Алоэ вера","Витамин Е","Масло ши"],
-    "550 ₽"
-  ),
-  new Product( 
-    "img/product2.jpg",
-    "Тоник свежающий",
-    "Уход",
-    "Очищение и сужение пор",
-    ["Гамамелис", "Мята", "Зеленый чай"],
-    "420 ₽"
-  )
- ];
+
+const creamMousse = new Product(
+  "product1",
+  "Крем увлажняющий",
+  "Уход",
+  "Увлажнение и питание кожи",
+  ["Алоэ вера", "Витамин Е", "Масло ши"],
+  "550 ₽"
+);
+
+const tonik = new Product(
+  "product2",
+  "Тоник освежающий",
+  "Уход",
+  "Очищение и сужение пор",
+  ["Гамамелис", "Мята", "Зеленый чай"],
+  "420 ₽"
+);
+
+export const products = [creamMousse, tonik];
